requirements/quality.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
appdirs==1.4.4            # via -r requirements/test.txt, virtualenv
attrs==20.3.0             # via -r requirements/test.txt, pytest
coverage==5.3.1           # via -r requirements/test.txt, pytest-cov
ddt==1.4.1                # via -r requirements/test.txt
distlib==0.3.1            # via -r requirements/test.txt, virtualenv
django-waffle==2.0.0      # via -r requirements/test.txt
django==2.2.17            # via -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt, -r requirements/test.txt
filelock==3.0.12          # via -r requirements/test.txt, tox, virtualenv
iniconfig==1.1.1          # via -r requirements/test.txt, pytest
isort==5.7.0              # via -r requirements/quality.in
mock==4.0.3               # via -r requirements/test.txt
packaging==20.8           # via -r requirements/test.txt, pytest, tox
path.py==12.5.0           # via -r requirements/test.txt
path==15.0.1              # via -r requirements/test.txt, path.py
pluggy==0.13.1            # via -r requirements/test.txt, pytest, tox
py==1.10.0                # via -r requirements/test.txt, pytest, tox
pycodestyle==2.6.0        # via -r requirements/quality.in
pydocstyle==5.1.1         # via -r requirements/quality.in
pyparsing==2.4.7          # via -r requirements/test.txt, packaging
pytest-cov==2.11.0        # via -r requirements/test.txt
pytest-django==4.1.0      # via -r requirements/test.txt
pytest==6.2.1             # via -r requirements/test.txt, pytest-cov, pytest-django
pytz==2020.5              # via -r requirements/test.txt, django
pyyaml==5.3.1             # via -r requirements/test.txt
six==1.15.0               # via -r requirements/test.txt, tox, virtualenv
snowballstemmer==2.0.0    # via pydocstyle
sqlparse==0.4.1           # via -r requirements/test.txt, django
toml==0.10.2              # via -r requirements/test.txt, pytest, tox
tox==3.21.2               # via -r requirements/test.txt
virtualenv==20.4.0        # via -r requirements/test.txt, tox
